@@ -13,8 +13,8 @@
         }
     },
     "require": {
-        "php": "^7.1",
-        "php-di/php-di": "6.0.0-beta3",
+        "php": "^7.4",
+        "php-di/php-di": "^6.0",
         "stratify/framework": "~0.7.1",
         "stratify/twig-module": "~0.3.0",
         "twig/twig": "~1.35.0",
@@ -26,7 +26,7 @@
         "symfony/monolog-bridge": "^3.1",
         "symfony/event-dispatcher": "^3.1",
         "psr7-sessions/storageless": "^2.0",
-        "league/commonmark": "^0.14.0",
+        "league/commonmark": "^1.3",
         "gravatarphp/gravatar": "^1.0",
         "gravatarphp/twig-integration": "^1.0",
         "league/oauth2-client": "^2.0",
@@ -37,11 +37,11 @@
         "zbateson/mail-mime-parser": "^0.4.3",
         "vlucas/phpdotenv": "^2.4",
         "sentry/sentry": "^1.7",
-        "bref/bref": "~0.4.0",
+        "bref/bref": "~0.5.17",
         "bref/logger": "^1.0.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.0",
+        "phpunit/phpunit": "^7.0",
         "mnapoli/hard-mode": "^0.1.1"
     }
 }
